Add UI BaseUrlFragment
