+++ app/src/main/java/com/ngxqt/mdm/ui/fragments/BaseUrlFragment.kt
@@ -6,6 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.activity.addCallback
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
@@ -15,6 +16,7 @@ import com.ngxqt.mdm.R
 import com.ngxqt.mdm.data.local.UserPreferences
 import com.ngxqt.mdm.databinding.FragmentBaseUrlBinding
 import com.ngxqt.mdm.ui.viewmodels.BaseUrlViewModel
+import com.ngxqt.mdm.util.isUrlValid
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
 
@@ -23,6 +25,7 @@ class BaseUrlFragment : Fragment() {
     private val  viewModel: BaseUrlViewModel by viewModels()
     private var _binding: FragmentBaseUrlBinding? = null
     private val binding get() = _binding!!
+    private var backPressedCount = 0
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -37,16 +40,36 @@ class BaseUrlFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         Log.d("BaseUrlFragment","onViewCreated")
+        setBackPressed()
+        binding.buttonSaveUrl.setOnClickListener {
+            val baseUrl = binding.editTextBaseUrl.text.toString().trim()
+            if (baseUrl.isNotEmpty() && isUrlValid(baseUrl)) saveBaseUrl(baseUrl)
+            else Toast.makeText(requireContext(),"Địa Chỉ Trang Web Không Đúng", Toast.LENGTH_SHORT).show()
+        }
+    }
 
+    private fun saveBaseUrl(baseUrl: String) {
         lifecycleScope.launch {
-            val saveDeferred = viewModel.saveBaseUrl("https://bvkienanhp.qltbyt.com")
+            val saveDeferred = viewModel.saveBaseUrl(baseUrl)
             saveDeferred.await()
-            // Khởi tạo lại Api Module ở đây để Retrofit cập nhật giá trị mới của baseUrl
             Log.d("BaseUrlFragment", "RUN HERE")
             UserPreferences(requireContext()).accessBaseUrlString().let {
                 Log.d("BaseUrlFragment", "token ${it}")
                 if (it.isNullOrEmpty()) Toast.makeText(requireContext(),"BaseURL is ${it}", Toast.LENGTH_SHORT).show()
                 else findNavController().navigate(R.id.action_baseUrlFragment5_to_loginFragment)
+            }
+        }
+    }
+
+
+
+    private fun setBackPressed(){
+        requireActivity().onBackPressedDispatcher.addCallback(viewLifecycleOwner) {
+            if (backPressedCount == 1){
+                requireActivity().finishAffinity()
+            } else {
+                backPressedCount++
+                Toast.makeText(requireContext(), "Nhấn hai lần để thoát ứng dụng", Toast.LENGTH_SHORT).show()
             }
         }
     }
