+++ app/src/main/java/com/ngxqt/mdm/util/Utils.kt
@@ -1,0 +1,10 @@
+package com.ngxqt.mdm.util
+
+internal fun isEmailValid(email: String): Boolean {
+    return android.util.Patterns.EMAIL_ADDRESS.matcher(email).matches()
+}
+
+internal fun isUrlValid(url: String): Boolean {
+    val regex = Regex("^(http|https)://[\\w.-]+\\.com$")
+    return regex.matches(url)
+}
