+++ app/src/main/res/drawable/ic_web.xml
@@ -1,0 +1,4 @@
+<vector android:height="24dp" android:viewportHeight="24"
+    android:viewportWidth="24" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#FF000000" android:pathData="M12,0A12,12 0,1 0,24 12,12.013 12.013,0 0,0 12,0ZM20.657,7L17.431,7A19.675,19.675 0,0 0,14.6 2.342,10.041 10.041,0 0,1 20.657,7ZM16.5,12a10.209,10.209 0,0 1,-0.473 3L7.973,15A10.209,10.209 0,0 1,7.5 12a10.209,10.209 0,0 1,0.473 -3h8.054A10.209,10.209 0,0 1,16.5 12ZM8.774,17h6.452A19.562,19.562 0,0 1,12 21.588,19.562 19.562,0 0,1 8.774,17ZM8.774,7A19.562,19.562 0,0 1,12 2.412,19.562 19.562,0 0,1 15.226,7ZM9.4,2.342A19.7,19.7 0,0 0,6.569 7L3.343,7A10.041,10.041 0,0 1,9.4 2.342ZM2.46,9L5.892,9a11.676,11.676 0,0 0,0 6L2.46,15a10.013,10.013 0,0 1,0 -6ZM3.343,17L6.569,17A19.675,19.675 0,0 0,9.4 21.658,10.041 10.041,0 0,1 3.343,17ZM14.6,21.658A19.7,19.7 0,0 0,17.431 17h3.226A10.041,10.041 0,0 1,14.6 21.658ZM21.54,15L18.108,15a11.676,11.676 0,0 0,0 -6L21.54,9a10.013,10.013 0,0 1,0 6Z"/>
+</vector>
