+++ app/src/main/res/layout/fragment_base_url.xml
@@ -1,13 +1,95 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:paddingLeft="24dp"
+    android:paddingRight="30dp"
+    android:background="@color/white"
     tools:context=".ui.fragments.BaseUrlFragment">
 
     <TextView
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="Hello" />
+        android:layout_height="wrap_content"
+        android:fontFamily="@font/averta_std_cy_bold_1"
+        android:gravity="center"
+        android:text="Ứng Dụng Quản Lý Thiết Bị Y Tế"
+        android:textColor="@color/black"
+        android:textSize="18sp"
+        app:layout_constraintBottom_toTopOf="@+id/image_cardView"
+        app:layout_constraintTop_toTopOf="@+id/imageViewLogo"
+        app:layout_constraintVertical_bias="0.878"
+        tools:layout_editor_absoluteX="24dp" />
 
-</FrameLayout>
+    <ImageView
+        android:id="@+id/imageViewLogo"
+        android:layout_width="250dp"
+        android:layout_height="250dp"
+        android:layout_marginBottom="32dp"
+        app:layout_constraintBottom_toTopOf="@+id/image_cardView"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.496"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.601"
+        app:srcCompat="@drawable/logo" />
+
+    <androidx.cardview.widget.CardView
+        android:id="@+id/image_cardView"
+        android:layout_width="match_parent"
+        android:layout_height="56sp"
+        android:backgroundTint="@color/edit_text_color"
+        app:cardCornerRadius="20dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.444"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        >
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+
+            <EditText
+                android:id="@+id/editTextBaseUrl"
+                android:layout_width="0dp"
+                android:layout_height="match_parent"
+                android:background="@drawable/edit_text_background"
+                android:drawableLeft="@drawable/ic_web"
+                android:drawablePadding="10.24dp"
+                android:ems="10"
+                android:fontFamily="@font/averta_std_cy_bold_1"
+                android:hint="Nhập địa chỉ trang web của bệnh viện"
+                android:text="http://bvdemo.qltbyt.com"
+                android:paddingLeft="21.76dp"
+                android:textColor="@color/black50"
+                android:textSize="14sp"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                tools:ignore="SpeakableTextPresentCheck,TouchTargetSizeCheck" />
+        </androidx.constraintlayout.widget.ConstraintLayout>
+    </androidx.cardview.widget.CardView>
+
+    <Button
+        android:id="@+id/buttonSaveUrl"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="14dp"
+        android:padding="17dp"
+        android:background="@drawable/button_login"
+        android:fontFamily="@font/averta_std_cy_semibold_1"
+        android:letterSpacing="0.001"
+        android:text="BẮT ĐẦU"
+        android:textColor="@color/white"
+        android:textSize="16sp"
+        android:textStyle="bold"
+        app:backgroundTint="@null"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.444"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/image_cardView"
+        app:layout_constraintVertical_bias="0.017" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
